Changed module entry
- Updated module entry from  'nuxt-kudra/module' to 'nuxt-kudra'
- Updated documentation to reflect these changes
- Updated the package version to 1.1.0
