+++ src/index.ts
@@ -1,3 +1,5 @@
+import { Module } from "@nuxt/types";
+import { Kudra } from "./kudra";
 import { META_NAME } from "./meta";
 import { KudraOptions } from "./options";
 
@@ -9,3 +11,14 @@ declare module "@nuxt/types" {
 
 export { meta } from "./meta";
 export * from "./runtime";
+
+const kudraModule: Module<KudraOptions> = function (moduleOptions) {
+  /* istanbul ignore next */
+  if (!this.options.dev) {
+    this.options.build.transpile?.push(META_NAME);
+  }
+
+  new Kudra(this, moduleOptions);
+};
+
+export default kudraModule;
